backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
# later: openai